common: logging: Restructure backend code
