+++ src/common/logging/backend.h
@@ -1,9 +1,9 @@
 // Copyright 2014 Citra Emulator Project
 // Licensed under GPLv2 or any later version
 // Refer to the license.txt file included.
+
 #pragma once
 
-#include <chrono>
 #include <filesystem>
 #include <memory>
 #include <string>
@@ -18,21 +18,6 @@ class IOFile;
 namespace Common::Log {
 
 class Filter;
-
-/**
- * A log entry. Log entries are store in a structured format to permit more varied output
- * formatting on different frontends, as well as facilitating filtering and aggregation.
- */
-struct Entry {
-    std::chrono::microseconds timestamp;
-    Class log_class{};
-    Level log_level{};
-    const char* filename = nullptr;
-    unsigned int line_num = 0;
-    std::string function;
-    std::string message;
-    bool final_entry = false;
-};
 
 /**
  * Interface for logging backends. As loggers can be created and removed at runtime, this can be
@@ -129,17 +114,6 @@ void AddBackend(std::unique_ptr<Backend> backend);
 void RemoveBackend(std::string_view backend_name);
 
 Backend* GetBackend(std::string_view backend_name);
-
-/**
- * Returns the name of the passed log class as a C-string. Subclasses are separated by periods
- * instead of underscores as in the enumeration.
- */
-const char* GetLogClassName(Class log_class);
-
-/**
- * Returns the name of the passed log level as a C-string.
- */
-const char* GetLevelName(Level log_level);
 
 /**
  * The global filter will prevent any messages from even being processed if they are filtered. Each
